backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "tunehound"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.135",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.14",
    "aiosqlite>=0.20",
    "yt-dlp>=2025.1",
    "mutagen>=1.47",
    "musicbrainzngs>=0.7",
    "httpx>=0.28",
    "watchdog>=6.0",
    "pydantic-settings>=2.13",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.28",
    "anyio>=4.0",
    "ruff>=0.9",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (formatter handles this)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F401"]  # unused imports in test files are fine (re-exports)
